docs(storybook): update pagination to storybook 5
